+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.2
+
+- Checkbox input element
+
 ## 0.0.1
 
-- Initial version, created by Stagehand
+- Basic elements: Div, Text, Image
+- ItemMaker for children